+++ libraries/effect/src/androidTest/java/androidx/media3/effect/DefaultVideoFrameProcessorImageFrameOutputTest.java
@@ -16,7 +16,6 @@
 package androidx.media3.effect;
 
 import static androidx.media3.common.util.Assertions.checkNotNull;
-import static androidx.media3.common.util.TimestampIterator.createFromLongIterator;
 import static androidx.media3.test.utils.BitmapPixelTestUtil.readBitmap;
 import static com.google.common.truth.Truth.assertThat;
 
@@ -24,9 +23,12 @@ import android.graphics.Bitmap;
 import android.util.Pair;
 import androidx.media3.common.C;
 import androidx.media3.common.ColorInfo;
+import androidx.media3.common.util.TimestampIterator;
 import androidx.media3.test.utils.VideoFrameProcessorTestRunner;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.common.collect.ImmutableList;
+import java.util.Iterator;
+import java.util.List;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -210,12 +212,34 @@ public class DefaultVideoFrameProcessorImageFrameOutputTest {
     videoFrameProcessorTestRunner.queueInputBitmaps(
         bitmap1.getWidth(),
         bitmap1.getHeight(),
-        Pair.create(bitmap1, createFromLongIterator(ImmutableList.of(offset1).iterator())),
-        Pair.create(
-            bitmap2, createFromLongIterator(ImmutableList.of(offset2, offset3).iterator())));
+        Pair.create(bitmap1, createTimestampIterator(ImmutableList.of(offset1))),
+        Pair.create(bitmap2, createTimestampIterator(ImmutableList.of(offset2, offset3))));
     videoFrameProcessorTestRunner.endFrameProcessing();
 
     assertThat(actualPresentationTimesUs).containsExactly(offset1, offset2, offset3).inOrder();
+  }
+
+  private static TimestampIterator createTimestampIterator(List<Long> elements) {
+
+    Iterator<Long> elementsIterator = elements.iterator();
+
+    return new TimestampIterator() {
+      @Override
+      public boolean hasNext() {
+        return elementsIterator.hasNext();
+      }
+
+      @Override
+      public long next() {
+        return elementsIterator.next();
+      }
+
+      @Override
+      public TimestampIterator copyOf() {
+        // Method not needed for effects tests.
+        throw new UnsupportedOperationException();
+      }
+    };
   }
 
   private VideoFrameProcessorTestRunner.Builder getDefaultFrameProcessorTestRunnerBuilder(