+++ libraries/transformer/src/main/java/androidx/media3/transformer/SampleConsumer.java
@@ -72,24 +72,6 @@ public interface SampleConsumer {
    * <p>Should only be used for image data.
    *
    * @param inputBitmap The {@link Bitmap} to queue to the consumer.
-   * @param durationUs The duration for which to display the {@code inputBitmap}, in microseconds.
-   * @param frameRate The frame rate at which to display the {@code inputBitmap}, in frames per
-   *     second.
-   * @return Whether the {@link Bitmap} was successfully queued. If {@code false}, the caller should
-   *     try again later.
-   */
-  // TODO(b/262693274): Delete this method and usages in favor of the one below (Note it is not
-  //   deprecated because transformer still relies on this method for frame duplication).
-  default boolean queueInputBitmap(Bitmap inputBitmap, long durationUs, int frameRate) {
-    throw new UnsupportedOperationException();
-  }
-
-  /**
-   * Attempts to provide an input {@link Bitmap} to the consumer.
-   *
-   * <p>Should only be used for image data.
-   *
-   * @param inputBitmap The {@link Bitmap} to queue to the consumer.
    * @param inStreamOffsetsUs The times within the current stream that the bitmap should be
    *     displayed at. The timestamps should be monotonically increasing.
    */